backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
openai==1.55.0
python-dotenv==1.0.1
pydantic==2.10.2
